Fix NullQuery statements
