+++ decidim-core/lib/decidim/query.rb
@@ -6,11 +6,21 @@
 # We have done this so we can decouple Decidim from any Virtus dependency, which is a dead project
 # Please follow Decidim discussion to understand more https://github.com/decidim/decidim/discussions/7234
 module Decidim
+  class UnableToComposeQueriesException < StandardError
+    def initialize(query, other)
+      super(
+        "Unable to composite queries #{query.class.name} and " \
+        "#{other.class.name}. You cannot compose queries where #query " \
+        "returns an ActiveRecord::Relation in one and an array in the other."
+      )
+    end
+  end
+
   class Query
     include Enumerable
 
     def self.merge(*queries)
-      queries.reduce(Decidim::Query.new(nil)) { |a, e| a.merge(e) }
+      queries.reduce(Decidim::Query.new([])) { |a, e| a.merge(e) }
     end
 
     def initialize(scope = ActiveRecord::NullRelation)
@@ -18,13 +28,11 @@ module Decidim
     end
 
     def query
-      return [] if @scope.nil?
-
       @scope
     end
 
     def |(other)
-      return Decidim::Query.new(other.cached_query) if @scope.nil?
+      return other if @scope.is_a?(Array) && @scope == []
 
       if relation? && other.relation?
         Decidim::Query.new(cached_query.merge(other.cached_query))
